Add files via upload
This script disables users and computers that have not logged in for 30 days, moves them to an OUs titled "Disabled", and adds a note in the description of each account where the user's/computer's original OU was.  It will also scan the disabled OUs for accounts that have not logged in for 60 days or more, write them to a CSV file for record keeping that is exported to the C:\, and then delete them permanently.

This script has descriptions above each action and annotates where changes must be made for your network.  It was written based on the DoD standards for account activity.

###Use at Your Own Risk###