Update Pull Device Information from single device or OU.ps1
requires user input in various stages on the script prompting for credentials and also the scope of the query (entire domain or single machine).  Places all info pulled from machines and places info in an excel spreadsheet categorized in the tabs.  makes life so much easier for that junior admin trying to validate a capstone document or a property book/Hand receipt.